requirements.txt
langchain
langchain-core

langchain-openai
openai

langchain-anthropic

langchain-google-genai

langchain-huggingface
transformers
huggingface_hub

python-dotenv

numpy
scikit-learn